+++ test/videodownlinkbandwidthpolicy/VideoPriorityBasedPolicy.test.ts
@@ -4,7 +4,6 @@
 import * as chai from 'chai';
 import * as sinon from 'sinon';
 
-import { ServerSideNetworkAdaption, VideoPriorityBasedPolicyConfig } from '../../src';
 import AudioVideoTileController from '../../src/audiovideocontroller/AudioVideoController';
 import NoOpAudioVideoTileController from '../../src/audiovideocontroller/NoOpAudioVideoController';
 import ClientMetricReport from '../../src/clientmetricreport/ClientMetricReport';
@@ -14,6 +13,7 @@ import GlobalMetricReport from '../../src/clientmetricreport/GlobalMetricReport'
 import StreamMetricReport from '../../src/clientmetricreport/StreamMetricReport';
 import ContentShareConstants from '../../src/contentsharecontroller/ContentShareConstants';
 import NoOpDebugLogger from '../../src/logger/NoOpDebugLogger';
+import ServerSideNetworkAdaption from '../../src/signalingclient/ServerSideNetworkAdaption';
 import {
   SdkBitrate,
   SdkBitrateFrame,
@@ -27,9 +27,11 @@ import VideoDownlinkObserver from '../../src/videodownlinkbandwidthpolicy/VideoD
 import VideoPreference from '../../src/videodownlinkbandwidthpolicy/VideoPreference';
 import { VideoPreferences } from '../../src/videodownlinkbandwidthpolicy/VideoPreferences';
 import VideoPriorityBasedPolicy from '../../src/videodownlinkbandwidthpolicy/VideoPriorityBasedPolicy';
+import VideoPriorityBasedPolicyConfig from '../../src/videodownlinkbandwidthpolicy/VideoPriorityBasedPolicyConfig';
 import VideoQualityAdaptationPreference from '../../src/videodownlinkbandwidthpolicy/VideoQualityAdaptationPreference';
 import SimulcastVideoStreamIndex from '../../src/videostreamindex/SimulcastVideoStreamIndex';
 import VideoTileController from '../../src/videotilecontroller/VideoTileController';
+import DOMMockBehavior from '../dommock/DOMMockBehavior';
 import DOMMockBuilder from '../dommock/DOMMockBuilder';
 
 describe('VideoPriorityBasedPolicy', () => {
@@ -41,6 +43,8 @@ describe('VideoPriorityBasedPolicy', () => {
   let audioVideoController: AudioVideoTileController;
   let tileController: VideoTileController;
 
+  let domMockBuilder: DOMMockBuilder;
+  let behavior: DOMMockBehavior;
   interface DateNow {
     (): number;
   }
@@ -300,9 +304,14 @@ describe('VideoPriorityBasedPolicy', () => {
     startTime = Date.now();
     originalDateNow = Date.now;
     Date.now = mockDateNow;
+    behavior = new DOMMockBehavior();
+    domMockBuilder = new DOMMockBuilder(behavior);
     audioVideoController = new NoOpAudioVideoTileController();
     tileController = audioVideoController.videoTileController;
-    policy = new VideoPriorityBasedPolicy(logger);
+    const policyConfig = new VideoPriorityBasedPolicyConfig();
+    // Most of the tests below are for the legacy path without server side network adaptation
+    policyConfig.serverSideNetworkAdaption = ServerSideNetworkAdaption.None;
+    policy = new VideoPriorityBasedPolicy(logger, policyConfig);
     policy.bindToTileController(tileController);
     videoStreamIndex = new SimulcastVideoStreamIndex(logger);
   });
@@ -336,7 +345,6 @@ describe('VideoPriorityBasedPolicy', () => {
 
   describe('worksWithoutTileController', () => {
     it('runs without tile controller', () => {
-      const policy = new VideoPriorityBasedPolicy(logger);
       updateIndexFrame(videoStreamIndex, 6, 0, 600);
       policy.updateIndex(videoStreamIndex);
       const preferences = VideoPreferences.prepare();
@@ -351,7 +359,6 @@ describe('VideoPriorityBasedPolicy', () => {
 
   describe('default preference', () => {
     it('use default preference', () => {
-      const policy = new VideoPriorityBasedPolicy(logger);
       updateIndexFrame(videoStreamIndex, 1, 0, 1200);
       policy.updateIndex(videoStreamIndex);
       let resub = policy.wantsResubscribe();
@@ -389,8 +396,9 @@ describe('VideoPriorityBasedPolicy', () => {
     });
 
     it('use default preference with bandwidth probing', () => {
-      const policy = new VideoPriorityBasedPolicy(logger);
-      policy.setServerSideNetworkAdaption(ServerSideNetworkAdaption.BandwidthProbing);
+      const config = new VideoPriorityBasedPolicyConfig();
+      config.serverSideNetworkAdaption = ServerSideNetworkAdaption.BandwidthProbing;
+      const policy = new VideoPriorityBasedPolicy(logger, config);
 
       updateIndexFrame(videoStreamIndex, 1, 0, 1200);
       policy.updateIndex(videoStreamIndex);
@@ -429,7 +437,6 @@ describe('VideoPriorityBasedPolicy', () => {
     });
 
     it('pause tiles if not enough bandwidth for default preference', () => {
-      const policy = new VideoPriorityBasedPolicy(logger);
       policy.bindToTileController(tileController);
       updateIndexFrame(videoStreamIndex, 3, 0, 1000);
       policy.updateIndex(videoStreamIndex);
@@ -628,6 +635,25 @@ describe('VideoPriorityBasedPolicy', () => {
 
       // @ts-ignore
       expect(policy.videoPreferencesUpdated).to.be.false;
+
+      updateIndexFrame(videoStreamIndex, 3, 0, 600);
+      policy.updateIndex(videoStreamIndex);
+
+      const resub = policy.wantsResubscribe();
+      expect(resub).to.equal(true);
+      const received = policy.chooseSubscriptions();
+      expect(received.array()).to.deep.equal([2, 4, 6]);
+    });
+
+    it('Interprets default value as BandwidthProbingAndRemoteVideoQualityAdaption', async () => {
+      const config = new VideoPriorityBasedPolicyConfig();
+      config.serverSideNetworkAdaption = ServerSideNetworkAdaption.Default;
+      const policy = new VideoPriorityBasedPolicy(logger, config);
+
+      // @ts-ignore
+      expect(policy.videoPriorityBasedPolicyConfig.serverSideNetworkAdaption).to.be.eq(
+        ServerSideNetworkAdaption.BandwidthProbingAndRemoteVideoQualityAdaption
+      );
 
       updateIndexFrame(videoStreamIndex, 3, 0, 600);
       policy.updateIndex(videoStreamIndex);
@@ -972,7 +998,9 @@ describe('VideoPriorityBasedPolicy', () => {
 
     it('Probe fail with StableNetworkPreset', () => {
       updateIndexFrame(videoStreamIndex, 4, 300, 1200);
-      policy.setVideoPriorityBasedPolicyConfigs(VideoPriorityBasedPolicyConfig.StableNetworkPreset);
+      const policyConfig = VideoPriorityBasedPolicyConfig.StableNetworkPreset;
+      policyConfig.serverSideNetworkAdaption = ServerSideNetworkAdaption.None;
+      policy.setVideoPriorityBasedPolicyConfigs(policyConfig);
       policy.updateIndex(videoStreamIndex);
       const preferences = VideoPreferences.prepare();
       preferences.add(new VideoPreference('attendee-1', 2));
@@ -1070,7 +1098,6 @@ describe('VideoPriorityBasedPolicy', () => {
 
   describe('paused', () => {
     it('Tile added but not in subscribe', async () => {
-      const domMockBuilder = new DOMMockBuilder();
       const observer: VideoDownlinkObserver = {
         tileWillBePausedByDownlinkPolicy(_tileId: number) {},
         tileWillBeUnpausedByDownlinkPolicy(_tileId: number) {},
@@ -1502,8 +1529,9 @@ describe('VideoPriorityBasedPolicy', () => {
 
   describe('VideoPriorityBasedPolicyConfig', () => {
     it('will not instantly drop videos caused by dip during startup period', () => {
-      const policy = new VideoPriorityBasedPolicy(logger);
-      policy.setVideoPriorityBasedPolicyConfigs(VideoPriorityBasedPolicyConfig.StableNetworkPreset);
+      const policyConfig = VideoPriorityBasedPolicyConfig.StableNetworkPreset;
+      policyConfig.serverSideNetworkAdaption = ServerSideNetworkAdaption.None;
+      policy.setVideoPriorityBasedPolicyConfigs(policyConfig);
       updateIndexFrame(videoStreamIndex, 1, 0, 1200);
       policy.updateIndex(videoStreamIndex);
       let resub = policy.wantsResubscribe();
@@ -1529,8 +1557,9 @@ describe('VideoPriorityBasedPolicy', () => {
     });
 
     it('unstable network with unstable preset', () => {
-      const config = VideoPriorityBasedPolicyConfig.UnstableNetworkPreset;
-      policy.setVideoPriorityBasedPolicyConfigs(config);
+      const policyConfig = VideoPriorityBasedPolicyConfig.UnstableNetworkPreset;
+      policyConfig.serverSideNetworkAdaption = ServerSideNetworkAdaption.None;
+      policy.setVideoPriorityBasedPolicyConfigs(policyConfig);
       updateIndexFrame(videoStreamIndex, 3, 300, 1200);
       policy.updateIndex(videoStreamIndex);
       const metricReport = new ClientMetricReport(logger);
@@ -1613,6 +1642,7 @@ describe('VideoPriorityBasedPolicy', () => {
 
     it('stable network with stable preset', () => {
       const config = VideoPriorityBasedPolicyConfig.StableNetworkPreset;
+      config.serverSideNetworkAdaption = ServerSideNetworkAdaption.None;
       policy.setVideoPriorityBasedPolicyConfigs(config);
       updateIndexFrame(videoStreamIndex, 3, 300, 1200);
       policy.updateIndex(videoStreamIndex);
@@ -1734,7 +1764,6 @@ describe('VideoPriorityBasedPolicy', () => {
     });
 
     it('works with non goog stats', () => {
-      const policy = new VideoPriorityBasedPolicy(logger);
       updateIndexFrame(videoStreamIndex, 1, 0, 1200);
       policy.updateIndex(videoStreamIndex);
       let resub = policy.wantsResubscribe();